Added function to get new data from database
