+++ src/Components/Dashboard/Clusters/Clusters.js
@@ -1,10 +1,9 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Button, makeStyles, Container } from '@material-ui/core';
 import { CreateCluster } from './CreateCluster';
 import { serverURL } from '../Dashboard';
 import { useDispatch, useSelector } from 'react-redux';
 import { showVMGroup } from '../../../Actions/VMGroupActions';
-import { showClusters } from '../../../Actions/ClusterActions'
 import { useRouteMatch, Link, Switch, Route } from 'react-router-dom';
 import { TableOfClusters } from './TableOfClusters';
 
@@ -20,25 +19,13 @@ export function Clusters () {
 	let { path, url } = useRouteMatch();
 	const dispatch = useDispatch();
 	const clusters = useSelector(state => state.clusters)
+
 	const refreshVMGroupData = () => {
 		fetch(`${serverURL}/api/proxmox/vm/group/list`)
 		.then(response => response.json())
 		.then(data => data.vm_group_list)
 		.then(data => dispatch(showVMGroup(data)));
 	};
-	const refreshClustersData = () => {
-		fetch(`${serverURL}/cluster/list`)
-		.then(response => response.json())
-		.then(data => data.kubernetes_cluster_list)
-		.then(data => dispatch(showClusters(data)))
-	}
-	useEffect(() => {
-		const interval = setInterval(() => {
-			refreshClustersData();
-		}, 4000);
-		return () => clearInterval(interval);
-	  }, []);
-
 	return(
 		<Container maxWidth="xl">
 			<Switch>
@@ -55,6 +42,7 @@ export function Clusters () {
 					</Link>
 					<TableOfClusters
 						clusters={clusters}
+						refreshVMGroupData={refreshVMGroupData}
 					/>
 				</Route>
 			</Switch>
